requirements.txt
black
gpiozero
pygame
smbus2
RPi.GPIO; sys_platform == 'linux'
PyYAML
python-dotenv